Front end piece with some code cleanup and readme updates
This add the front end bit (boy it's ugly) which is just some simple
JS/jQuery/HTML. Nothing too nice.

Also updates the README and cleans up a bit of now unused code. If
this were a real app I'd also recommend a bunch of ways to make this
code better, but for now it works and the goal of the task has been
accomplished.